Skip vagrant < 2.0 in which().
The which() function has code to skip anything that looks like a shell
script (starts with '#!') in order to avoid the gem version of vagrant.
Unfortunately the official vagrant package and the version shipped by
Fedora both use a shell script wrapper now, so this caused which() to
skip vagrant executables that should not have been skipped.

The gem version of vagrant was never updated past 1.5, so if we simply
ignore anything pre-2.0, we can safely skip the gem version as well as
old versions that may not work as expected.